Reuse code from subscribe module by extracting it to a trait.
